requirements.txt
beautifulsoup4>=4.12.2
lxml>=5.2.1
readability-lxml>=0.8.1
pdfminer.six>=20231228
pymupdf>=1.24.10
nltk>=3.9.1
spacy>=3.7.5
scikit-learn>=1.5.1
pandas>=2.2.2
numpy>=2.0.0
matplotlib>=3.9.0
pyyaml>=6.0.2
tqdm>=4.66.4
